Move speech command objects into a speech.commands module.
